// src/lib/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      reading: {
        Row: {
          chapter_id: string
          created_at: string
          id: string
          serial_number: number
          text: string
          translation: string
        }
        Insert: {
          chapter_id: string
          created_at?: string
          id: string
          serial_number: number
          text?: string
          translation?: string
        }
        Update: {
          chapter_id?: string
          created_at?: string
          id?: string
          serial_number?: number
          text?: string
          translation?: string
        }
        Relationships: [
          {
            foreignKeyName: "reading_chapter_id_fkey"
            columns: ["chapter_id"]
            isOneToOne: false
            referencedRelation: "chapters"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}